Refactor main into smaller modules
